requirements/base.txt
Django==1.9.7
Django-Select2==5.8.9
Pillow==3.4.1
Unidecode==0.4.19
cmsplugin-filer==1.1.3
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.1
django-filer==1.2.5
django-formtools==1.0
django-mptt==0.8.6
django-polymorphic==0.8.1
django-registration-redux==1.4
django-reversion==1.10.0
django-sekizai==0.10.0
django-treebeard==4.0.1
djangocms-admin-style==1.2.5
djangocms-attributes-field==0.1.1
djangocms-column==1.7.0
djangocms-inherit==0.2.2
djangocms-link==2.0.1
djangocms-maps==0.7.0
djangocms-text-ckeditor==3.3.0
djangorestframework==3.5.1
django-ical==1.4
easy-thumbnails==2.3
html5lib==0.9999999
pytz==2016.7
six==1.10.0
webencodings==0.5
aldryn-reversion==1.0.9
aldryn-style==1.1.1
aldryn-bootstrap3==1.1.2
aldryn-boilterplates==0.74
aldryn-newsblog==1.3.0
git+https://github.com/Chive/pyrate.git@feature/0.5.0#egg=pyrate
